dodao-lambdas/pull-tweets/pyproject.toml
[tool.poetry]
name = "pull-tweets"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
twscrape = "^0.14"
fastapi = "^0.115.5"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
python-dotenv = "^1.0.1"
boto3 = "^1.35.63"
botocore = "^1.35.63"
mangum = "^0.19.0"
pydantic = "^2.9.2"


[tool.poetry.group.dev.dependencies]
zappa = "^0.59.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
